backend/requirements-prod.txt
-r requirements.txt
gunicorn>=23.0.0
celery[redis]>=5.4.0
flower>=2.0.1
prometheus-fastapi-instrumentator>=7.0.0
opentelemetry-api>=1.36.0
opentelemetry-sdk>=1.36.0
opentelemetry-exporter-otlp>=1.36.0
opentelemetry-instrumentation-fastapi>=0.57b0
opentelemetry-instrumentation-logging>=0.57b0
opentelemetry-instrumentation-redis>=0.57b0
opentelemetry-instrumentation-requests>=0.57b0
opentelemetry-instrumentation-httpx>=0.57b0
redis[hiredis]>=6.2.0
psycopg[binary]>=3.2.0
sqlalchemy>=2.0.30
qdrant-client>=1.15.1
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=2.28.0
